feat: allow only authenticated users to book appointments
