+++ app/api/auth/[...nextauth]/route.ts
@@ -1,27 +1,6 @@
-import { db } from "@/app/_lib/prisma"
-import { PrismaAdapter } from "@auth/prisma-adapter"
+import { authOptions } from "@/app/_lib/auth"
 import NextAuth from "next-auth"
-import { Adapter } from "next-auth/adapters"
-import GoogleProvider from "next-auth/providers/google"
 
-const handler = NextAuth({
-  adapter: PrismaAdapter(db) as Adapter,
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-  ],
-  callbacks: {
-    async session({ session, user }) {
-      session.user = {
-        ...session.user,
-        id: user.id,
-      } as any
-      return session
-    },
-  },
-  secret: process.env.NEXT_AUTH_SECRET,
-})
+const handler = NextAuth(authOptions)
 
 export { handler as GET, handler as POST }
